refactor: streamline authentication components and introduce validation schemas for improved code organization and clarity
