+++ web/src/platform/layouts/Page.tsx
@@ -7,7 +7,7 @@ import { DevelopmentNotice } from '@/components/DevelopmentNotice';
 /** Renders public page chrome around page-specific content. */
 export default function Page() {
     return (
-        <Stack minHeight="100vh" gap={0}>
+        <Stack minHeight="100vh">
             <DevelopmentNotice />
             <Navbar />
             <Outlet />
