+++ web/src/platform/routes/auth/ForgotPassword.tsx
@@ -1,55 +1,23 @@
-import type { ReactNode } from 'react';
 import { z } from 'zod';
 import { Link } from '@astryxdesign/core/Link';
 import { Text } from '@astryxdesign/core/Text';
 import { Stack } from '@astryxdesign/core/Stack';
 import { Banner } from '@astryxdesign/core/Banner';
 import { Button } from '@astryxdesign/core/Button';
-import { Center } from '@astryxdesign/core/Center';
 import { useMutation } from '@tanstack/react-query';
-import { Heading } from '@astryxdesign/core/Heading';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Controller, useForm } from 'react-hook-form';
 import { TextInput } from '@astryxdesign/core/TextInput';
 import { api } from '@/lib/api';
 import { useToast } from '@/lib/hooks/use-toast';
+import { AuthLayout } from './AuthLayout';
+import { emailSchema } from './validation';
 
 const forgotPasswordSchema = z.object({
-    email: z.string().trim().min(1, 'Email is required').email('Enter a valid email address'),
+    email: emailSchema,
 });
 
 type ForgotPasswordValues = z.infer<typeof forgotPasswordSchema>;
-
-/** Renders password reset content within the standalone account page. */
-function AuthLayout({
-    children,
-    description,
-    title,
-}: {
-    children: ReactNode;
-    description: ReactNode;
-    title: ReactNode;
-}) {
-    return (
-        <Center minHeight="calc(100dvh - var(--appshell-header-height, 0px))" width="100%">
-            <Stack gap={4} maxWidth={384} paddingBlock={8} paddingInline={4} width="100%">
-                <Stack gap={1}>
-                    <Heading justify="center" level={1}>
-                        {title}
-                    </Heading>
-                    {typeof description === 'string' ? (
-                        <Text as="p" color="secondary" justify="center" type="supporting">
-                            {description}
-                        </Text>
-                    ) : (
-                        description
-                    )}
-                </Stack>
-                {children}
-            </Stack>
-        </Center>
-    );
-}
 
 /** Requests a password reset email without disclosing whether an account exists. */
 export default function ForgotPassword() {
