+++ web/src/platform/routes/docs/sdk/Index.tsx
@@ -15,7 +15,7 @@ function LocalRuntimeDiagram() {
                 <Card width="80%" variant="muted">
                     <Stack gap={3} align="center">
                         <UserRound aria-hidden className="text-accent" size={20} />
-                        <Stack gap={0} align="center">
+                        <Stack align="center">
                             <Text weight="semibold">User</Text>
                             <Text type="supporting">Browser</Text>
                         </Stack>
@@ -31,7 +31,7 @@ function LocalRuntimeDiagram() {
                 <Card width="80%" variant="muted">
                     <Stack gap={3} align="center">
                         <AppWindow aria-hidden className="text-accent" size={20} />
-                        <Stack gap={0} align="center">
+                        <Stack align="center">
                             <Text weight="semibold">Application</Text>
                             <Text type="supporting">Runtime</Text>
                         </Stack>
